requirements.txt
Flask==3.0.3
flasgger==0.9.7.1
gunicorn==22.0.0
python-dotenv==1.0.1
Werkzeug==3.0.4
ollama==0.3.3
PyMuPDF==1.24.9
python-docx==1.1.2
pandas==2.2.2
numpy==1.26.4
lxml==5.3.0
pillow==10.4.0
chardet==5.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pydantic==2.9.2
xlsxwriter==3.2.9
openpyxl==3.1.5